refact: implementado melhoria em rotas GET saldo e extrato
