+++ src/rotas.js
@@ -3,13 +3,11 @@
 const express = require('express');
 const rotas = express();
 const {listarContas, cadastrarConta, atualizarConta, deletarConta, depositar, sacar, transferir, saldo, extrato} = require('./controladores/controladores');
-const {validarSenhaBanco, validarEmailCpf, validarDados, validarContaBanco, validardeposito, validarSaque, validarTransferencia} = require('./intermediarios/intermediarios');
-
-
+const {validarSenhaBanco, validarEmailCpf, validarDados, validarContaBanco, validardeposito, validarSaque, validarTransferencia, validaSaldoExtrato} = require('./intermediarios/intermediarios');
 
 rotas.get('/contas', validarSenhaBanco , listarContas);
-rotas.get('/contas/saldo', saldo);
-rotas.get('/contas/extrato', extrato)
+rotas.get('/contas/saldo',validaSaldoExtrato, saldo);
+rotas.get('/contas/extrato',validaSaldoExtrato, extrato)
 
 rotas.post('/contas', validarEmailCpf, validarDados, cadastrarConta);
 rotas.post('/transacoes/depositar',validardeposito, depositar);
